Translate Types tests to Hspec, fix FIXMEs, find issues
I decided that we really should try using QuickCheck for running our
round trip testing. This is _mostly_ ok, but we have to fudge some stuff
for round-tripping lists via Haskell.

This currently fails tests for binary serialisation of bundles, which may
be due to bugs in the production code and needs further investigation.

NB: we've turned on compiler warnings just for these tests via the GHC_OPTIONS
    pragma. (I couldn't be bothered to work out by hand the imports I no longer
    used ;-) )