Change behaviour of serve'
The `doubleCatch`, which ended up inside the `doubleCatch` used by
`withListen`, wasn't really what we meant. We really just meant: "If we
terminate because the listen socket was closed, wait for children to
complete. If we terminate because we got killed, take down children."

This removes the strange "bad file descriptor" error on relay shutdown.

I also sneakily flipped the arguments to `throwAfter` as it seemed more
convenient to call it that way, but it doesn't seem as good
"grammatically".