Move xref validation outside of iterative value validation
because xrefs are non-local, and until we've traversed the tree updating
the types of all the nodes, we can't be sure that cross references will be
checked as being the type that they have been changed to.

NB: this is still broken, in that we also now need to track when xrefs
have been invalidated due to node type changes, because we're not re-
computing all the xrefs every time we validate the tree.