Change tack on how "post" data should arrive
Instead of taking it as a distinct new kind of message in a bundle, we will
take it as part of the container update stuff.

Note that we thoroughly ignore anything to do with it right now and that
whilst we have renamed PresentAfter to MoveAfter we haven't yet changed any
functionality!