Mirror Namespace Messages changes in Digests
Namely that subscriptions are now held as a Map of Sets with keys
corresponding to the Namespace. We remove SubOp and instead have separate
data structures for subscriptions and unsubscriptions, because we keep
wanting to query the data in that form anyway.

Because of the new data structures, we don't need qualify/unqualify helpers
any more either.