Break out Decodable from Encodable
The original intent of having `builder` and `parser` glued together was
that we couldn't ever define serialisation for something we then couldn't
deserialise. However, we actually want to start being asymmetric now. With
GADTs/DataKinds et al we might want to encode an explicitly typed value,
but can only decode an existential.