Redefine Definition to be a parameterised GADT
It's very much what we were trying to do with the individual definitions
and then the sum type wrapper. Now instead we have a single Definition
type, qualified by a lifted MetaType, and an existential wrapper
SomeDefinition for when we want to drop the type parameter.

We refine a DefName alias to make things more succinct and hide us from
the pain of changing underlying types if we ever do that again...