Add Util.Filterable and related functions
This began as an effort to get a more polymorphic filter function and to
confine some partially-defined functions like `fromJust` to a small enough
box where reasoning about them was trivial.

However, it grew into something slightly less trivial because I wanted to
be able to filter `Set`s with `justs`, `lefts` and `rights`. Hopefully,
the resulting library code isn't too nuts, and the use cases demonstrated
in the tests are compelling enough.