Break apart Valuespace errors using Wraps/Errs
so that we don't have one monolithic error and can differentiate between
errors that Providers and Consumers can raise.

Note, we do have to do some minor fiddling to get the errors all wrapped up
nicely, mostly in cases where we're not in the VsM' structure and not in a
simple MonadError.